Refactoring common input file functionality into a base class
This way applications do not have to drag all the generic baggage
if they do not need it.
